implement basic svlog lexer
Add svlog/lexer.rs that contains the starting point for a SystemVerilog
lexical analyzer. To this end, also add a Reader struct that allows for
buffered reads from a file. The reader shall convert the file to UTF8
encoding to simplify the lexer, and also provide a history of consumed
characters up to a certain point, usually the beginning of the token.

Symbol matching is implemented using a rust `match` expression, which
seems quite well-suited for this kind of task. Especially if the
optimizer is smart enough to rearrange the arms into a tree structure,
which seems likely given that LLVM is likely to reduce redundant
comparisons.