add basic parser generator
Add a basic parser generator in pre/pargen that derives a parser state
machine from a grammar. Make use of Rust's enums to stored the different
tokens and reduced symbols. The generated parser consists of a set of
action functions that handle the transition to the next state, including
proper reduction. External Rust functions are called to reduce the items
on the stack. There is still some room for improvement, but this is fine
as a starting point.