svlog: implement preprocessor
Implement a preprocessor for SystemVerilog. This is built around a
categorizing lexer roughly inspired by TeX, which assigns a category to
every input character and groups such characters up. This lexer emits a
stream of trivial text, number, symbol, and whitespace tokens that are
easier to digest in the preprocessor. The preprocessor then uses these
tokens to define, undefine, and substitute macros, but also to include
other files into the token stream.

Note that this should sooner or later be rewritten using one of the
parser combinator frameworks, which might be faster and provide more
maintainable code.