add destilled svlog expression grammar
Add the `docs/svlog_grammar.txt` file that contains a distilled version
of the grammar concerning expressions. Unfortunately the standard mixes
syntax and semantics in quite a few places, so the "official" syntax
needs to be broken down into something that can actually be parsed in a
reasonable way. Ambiguities need to be resolved at a later stage, when
more information about the parse is available.