svlog: pointless attempt at making ports useful
The syntax for ports is so convoluted that it seems almost impossible to
come up with a general parsing scheme. This whole thing is ambiguous to
such an incredible extent that the question arises whether it would not
be easier to have types be treated as a first class expression, with the
distinction between types and values being made only once the AST has
been built. Especially the combination of arbitrarily named types and
the possibility to have ports/declarations with an implicit type makes
distinguishing between a type signature `foo[42]` and a declaration name
`foo[42]` *very* hard. If types were first class citizens, both could be
parsed equally and the declaration would distinguish itself from the
type by the fact that it seemingly consists of two expressions: the type
and the declaration.