svlog: add basic type AST nodes
We need to build up the AST. Start with the implementation of the
`parse_data_type` function and modify it such that it produces a proper
AST node that contains the parsed information, along with the type's
span. Currently the packed dimensions that come with the type don't
carry any meaningful information. This needs to be changed as soon as
expressions can be parsed properly.