svlog: construct port AST nodes
Modify the code that parses ports and have it wrap the parsed data up in
a proper AST node. While doing this, apply the direction, type, and kind
inference rules outlined in the standard. I.e. inherit values from the
previous port or apply defaults.

Add a lifetime field to the module and interface AST nodes and have them
keep track of their port list.