svlog: support impl types in ports, param decls
Allow for implicit types in port lists. This is experimental still and
relies on converting the parsed type to a name. Better do this with a
parallel parser to try the implicit and explicit variant.

Implement parameter declaration parsing on the module level. This
ignores the potential type of the parameter for now, which is bad. Later
this should make use of the parallel parser to try the explicit and
implicit variant.