organize code into separate crates
Move all the src/*.rs files and the src/errors/mod.rs file to a separate
common crate. Then turn the src/svlog subdirectory into a separate
crate that relies on common. This should help speed up compilation times
and should allow for better unit testing of the individual parts of the
compiler.