svlog: implement port lowering to hir
Implement lowering of ports from the AST to HIR. This involves taking
the port lists present in the source text, detecting whether they use
ANSI or non-ANSI style, and gathering the necessary information to
arrive at a canonical port list. This last part consists of applying
default values for or carrying over from the previous port its port
kind, direction, and type, in the case of ANSI style; or going through
the module items and finding the port and var/net declarations that
go with a port, in the case of non-ANSI style.