test: add basic integration test
Add a `tests` directory that shall contain the integration tests for the
compiler. This mainly means HDL snippets that are compiled and checked
for various features. But the directory can also contain integration
tests in the rust sense, where we can test linking against the compiler
libraries.

For now add a parameter test that verifies some of the more involved
semantics of SystemVerilog parameters. They don't pass at the moment,
due to issues in the parser and then also further down the road.

Also add a `run.sh` file that runs all HDL integration tests through the
compiler. In the long run we might want to create a separate rust
executable which goes through the files, passes them to the appropriate
compiler, and then checks for certain properties as described in the
file's comments. But for now the shell script is likely to suffice.