svlog: improve error recovery within flanked
Improve the error recovery for parse errors that occur within a call to
`flanked`. Currently the parser simply aborts in between the opening and
closing delimiters if the inner function fails. This causes subsequent
calls to `recover_balanced` to fail, since the balancing of opening and
closing delimiters is broken. Alter `flanked` to recover properly to the
closing delimiter if the inner function fails. Also make unbalanced
delimiters a fatal error, since there is little chance of recovery.