svlog: improve recovery in item parsing
Improve the error recovery for items in the parser. The parser now fully
assumes that `parse_item` recovers properly to the end of an item. This
now moves the responsibility of recovering from arbitrary garbage on the
file root level into the `parse_item` function. Change that function to
recover to the next keyword that might introduce an item, e.g. `module`
or `interface`, in case some unexpected tokens are encountered.

Also improve the way `parse_class_decl` handles errors in the class
body. Until now the function just aborted upon once it recovered up to
the `endclass` keyword, but failing to consume that. Make the function
unconditionally consume the `endclass` keyword after recovery.