svlog: support type param assignment
Change the parameter assignment AST node such that it now accepts a new
`TypeOrExpr`, rather than a simple `Expr`. This allows type parameters
to be overridden. Note that the expression takes precedence, so there
are cases where a type would be parsed as an expression, e.g. when the
type is simply an identifier. This is exactly what the HIR is for: Find
these cases and resolve them, such that the MIR is unambiguous and
clean.