vhdl: implement lexer
Implement a VHDL lexer as per IEEE 1076-2008. The lexer makes use of a
new addition to the common library: `grind`. This is a module containing
a useful abstraction for parsers that can produce items when asked for
`next()`, and will emit any errors via a separate channel. Multiple
grinders chained together will cause errors at the front to propagate
backwards, such that long transformation chains can be established with
a well-defined point of error handling.