vhdl: lower lib to hir
Implement lowering design libraries to HIR. This implements the first
step in the layer-wise destructuring of the AST, where the tree is
transformed into individual nodes that reference other nodes via an ID.

Also implement determining the definitions within a library. Do this by
querying a library's HIR (which is generated on-demand if need be), then
iterating over the design units it contains and creating definitions as
needed. Soon based on this, architectures may be assigned to their
corresponding entity and package bodies to their corresponding package
declaration.

Also split out the `hir` module into a separate file.