remove use of @, clean up
Replace the use of `@matches` in `common/score.rs` with `MATCHES`. The
former worked on rustc v1.20, but fails on v1.19. Also clean up the
svlog part a bit, such that less warnings are produced during
compilation. At a later stage, the `#![allow(...)]` should be removed
and the issues properly fixed.