vhdl: impl entity lowering to hir
Implement the lowering of entities to HIR. Currently only the port and
generic clauses of the entity are considered. Declarations made within
the entity are ignored. These have to be added later on when the scopes
and definitions of the entity are to be built.

The lowering to HIR is the perfect location for additional semantic
checks that were omitted during parsing. For example, syntactically the
port and generic clauses are identical. The parsed AST is a superset of
what the VHDL grammar actually allows. While lowering, only the correct
parts of the parsed AST need to be retained, while the user can be
warned about the rest. The `node_ref_group!` macro allows the exact set
of acceptable nodes to be encoded in the type system.