vhdl: impl name resolution, split scoreboard
Split the `Scoreboard` types into `ScoreBoard` and `ScoreContext`. The
former holds all information that is being generated during the
compilation procedure, while the latter holds a reference to the
scoreboard and other context information. For the global scoreboard,
the context holds references to the `Session` and all scoreboards in
use (global, vhdl, svlog). For language-specific scoreboards, the
context holds references to the `Session`, the language scoreboard, and
the global scoreboard context through a `GenericContext` trait. This
allows the operations on the scoreboard to be detached from the
ownership of the scoreboard. Also, the context offers a location for
information that is necessary during processing, but is not required to
be embedded in the scoreboard. Moreover, the VHDL scoreboard needs a
reference to the global scoreboard, but the global scoreboard also needs
a reference to the VHDL scoreboard. This scheme takes care of this
dependency cycle.

Implement name resolution for VHDL. To this end, refine the `defs` table
of the scoreboard and add a `scope` table. The former holds all
definitions made in a scope, while the later also records the names made
visible e.g. through use clauses. Add a function to resolve names within
a scope. Also add a function to resolve compound names which returns the
resolved definitions and the rest of the name parts that were not
processed.

Add a preliminary means for declaring builtin libraries and packages.
Declare the `std` library and `standard`, `textio`, and `env` packages.
The packages are empty and do nothing useful as of now. Better implement
type and subtype declarations first.