vhdl: add type tests
Add test files for integer, enumerated, array, record, and access types.
These parse and elaborate at the moment, but only since the code
generation of architectures does not yet process signals. As soon as the
scoreboard tries to emit signals into the entity, these tests will start
to fail and require the corresponding types to be implemented.