vhdl: factor out definition gathering
Create a new `DefsContext` struct in `vhdl/defs.rs` that handles the
gathering of definitions in a context. The code in `vhdl/scope.rs` may
now create such a context and delegate the gathering of definitions to
it. Currently used by packages and architectures; other scopes are still
using the old approach.