vhdl: add signal assignment to hir
Add all forms of sequential signal assignment to the HIR. Lowering still
encounters problems because architectures try to immediately unpack
their statements. This causes a cyclic dependency, because this
unpacking requires names to be resolved, which in turn requires at least
a minimal amound of unpacking.

This can be fixed by not immediately unpacking statements when lowering
architectures, but rather postponing that until they are used. Pretty
much like the rest of the code until now.