vhdl: factor out typeck, add typeck trait
Move the typechecking into `vhdl/typeck.rs` and create a new `Typeck`
trait. The trait is implemented on the scoreboard context for all nodes
that do not directly have a type. This includes entities, architectures,
packages, statements, etc. The idea is that this trait triggers the
required type calculations, e.g. for the LHS and RHS of an assignment,
and then verifies that the types match up.