vhdl: lower array type declarations
Lower array type declarations to HIR. The array indices are not fully
interpreted yet, since the distinction between an unbounded index and an
index constrained by a subtype is quite difficult. We'll need more
helper functions in `lower_hir.rs` that allow names to be gradually and
carefully unpacked.

Array type declarations are not properly mapped to actual types from the
`ty` module yet. This is almost done, but requires some more
investigation as to what the requirements on indices are. More
specifically, indices need to be discrete ranges according to the
standard.

This contributes towards issues #25 and #35.