vhdl: impl array type
Add the `ArrayTy` type and map array type declarations to it. This is
merely a starting point and the implementation currently cannot emit the
proper LLHD type, nor the default value for an array. The former
requires functionality to determine the number of elements per array
dimension, the latter requires support in LLHD directly.

This contributes towards issues #25 and #35.

Also the implemented array pretty-printing fixes #34.