vhdl: refactor typeck and utils
Move `HasSpan` and `HasDesc` into a new `util` module in the common lib.
Refactor `typeck.rs` such that all type checking now operates on a
`TypeckContext` rather than the `ScoreContext`. This makes things easier
to handle and we can add more fine-grained functions to the context.
There is quite a proliferation of lifetimes which I don't know how to
tackle right now.