vhdl: add terms, proper support for array types
Implement the `Term`, which is a unified way of representing the
innermost constructs of the language. This includes primary names, the
operations on those names, as well as constraints applied to types. A
term's kind is determined not by the context in which it is used, but
rather the primary name and operations applied. This allows the lowering
to HIR pass to identify something as a "type", or a "subtype indication"
rather than having to figure these things out itself.