vhdl: fix infinite loop in defs, subtype via term
Fix an infinite loop that occurred when determining the definitions in
a scope. Type declarations required the HIR of the type in order to
declare enum variants, which caused an infinite loop since said type
would want to resolve names under certain circumstances. This is now
fixed.

Also, use termification to map all variants of subtypes to a `Term`,
then go back to a subtype via `term_to_subtype_ind`. Works quite well.