vhdl: termify aggregates and parentheses
Add the ability to convert aggregates and parentheses into terms. This
should now allow terms to be converted to constraints. Array constraints
can take interesting forms, e.g. `A (0 to 8) range 0 to 42`, where the
thing in parentheses is the index constraint, and `range ...` is the
element constraint. Getting the terms to properly group this up will be
quite interesting.