vhdl: avoid greedy subtype ind lowering
Fix a bug in VHDL where the compiler would get stuck in an infinite
recursion when lowering a signal to HIR. Since upon lowering the signal
the compiler issues a call to `unpack_subtype_ind` for the signal type,
and that function greedily lowers the subtype to HIR, and that process
involves name resolution, a circular dependency occurred.

We really need something in the compiler to avoid this, or at least to
point out the recursion and emit a bug.

This is now fixed with changing `unpack_subtype_ind` such that it does
not greedily lower to HIR, but rather stores the AST of the subtype to
be unpacked and registers a new ID for it.