vhdl: lower pkg and subprog bodies and insts
Implement proper lowering to HIR of package and subprogram declarations,
bodies, and instantiations. Also add an initial typeck implementation.
Some things are missing, especially the generic map and parameter port
list of subprograms.

This implements typeck of subprograms as per issue #43.
This implements lowering of subprograms to HIR as per issue #41.