vhdl: improve error msg for lazy nodes
Make the scoreboard track the spans for node IDs. Add a `bug` function
to the score context which can be used to report compiler bugs
associated with a node to the user. The function will augment the
message with the span stored for the node, if any. Have the lazy type
checking and evaluation report bugs via this new function.