vhdl: split aggregate into pos, named, others
In the code that maps a term to an aggregate, split the list of fields
into positional and named fields, and an optional `others`. Emit errors
appropriately. Doing this even before the aggregate goes to HIR is key
to make the rest of the pipeline lean, because otherwise each function
dealing with the aggregate would have to determine this information
again.