vhdl: impl part of aggregate typeval
Implement part of the type evaluation on aggregates. This is quite some
tricky business, since the appropriate mapping between the aggregate
fields and the record/array fields needs to be calculated first. Working
well for records, arrays are simply ignoring any multiple assignment as
of now. Maybe it's actually better to capture multiple assignment in
arrays further down the road, during constant evaluation or codegen.