vhdl: typeck const, sig, var, file decls
Add calls to `must_match` in the type checking phase for const, signal,
variable, and file declarations. This will now reject any initial values
for such declarations if they do not match the declaration's type.

The type checker is overzealous at the moment, complaining about pretty
much everything. This must be improved in the future. But too strict is
better than too lax.