vhdl: work towards new scoping mechanism
Add a new scope struct that will contain all defined names, as well as
imported names and scopes. Change the name resolution function such that
it takes this additional scoping mechanism into account. The existing
combination of `defs_table` and `scope_table` will be phased out in the
future. Scopes should be created and populated in the `add_*` functions.
The make context can help with that. Also, we need a proper root scope
that declares and imports the builtin packages, and a work scope which
is generated for each library individually, with `WORK` predefined.