vhdl: add builtin vector types
Add the following builtin types:

- `BOOLEAN_VECTOR`
- `BIT_VECTOR`
- `INTEGER_VECTOR`

To do this, named types have to be modified a bit. Rather than having
`Ty::Named` accept a `Span`, change it to accept a `TyName`. Then define
`TyName` to be either a `Span` or a `Name`. Usually types named in the
source code will have an associated span. Builtin types however do not
have a span, but must rather provide their name directly.