vhdl: add universal integer type
Add the universal integer type mandated for by the VHDL standard. All
abstract integer literals are of this type by default, and any value of
this type can implicitly converted into an integer type. This uncovered
a bug in the order how aggregate indices are type checked. Fix that as
well.