vhdl: add overload resolver
Add the `overload_resolver` module which is responsible for taking a set
of definitions and resolving them to one single definition that is
applicable given a certain type context. The resolution is split into
two steps: the reduction which eliminates all definitions that do not
apply, and the resolution which ensures that only one definition applies
after reduction.