vhdl: lookup ops as names and resolve
When generating unary and binary expression HIR nodes, convert the
operator to a resolvable name and lookup all definitions in the current
scope. Add these definitions to the HIR node.

Upon type evaluation, the overload resolver is used to filter the
available definitions and obtain an unambiguous implementation for the
operator. This does not yet work properly, since the builtin operators
register no types, which causes the `lazy_typeval` to fail. The next
step is to add such type information and complete the implementation of
the overload resolver.