vhdl: partially partially rework builtins
Change the way builtin types are handled. Rather than declaring bare
bones IDs all over the place and associating types and definitions with
them later, add a central `Builtin` and `BuiltinType` struct. These
encapsulate all the information of a builtin, and can be used to declare
them in a scope.