vhdl: define builtin operators along types
Migrate to a new means of defining builtin operators. These are a
byproduct of the definition of builtin types. Later on, also user
defined types will produce a series of implicitly defined builtin
operators, e.g. addition and subtraction for numeric types.

This takes things a step further. `test/vhdl/expr.vhd` now fails due to
the overload resolution not being implemented. This is quite nice and
easily fixable. There's also a few semantic errors in that file, e.g.
the use of shift operators on `INTEGER`. These need to be fixed in order
to get meaningful error messages.