vhdl: foundation of new type system
Add a new `vhdl::ty2` module which will be used to build up the new type
system. Once done the rest of the code will migrate from `ty` to `ty2`.
Eventually `ty2` will be renamed to `ty` and the original dropped.

This new approach makes better use of traits and enums. Basically each
type of VHDL is represented by a separate struct, e.g. `EnumType`. These
structs implement the `Type` trait which provides a common interface for
interacting with types. An additional enum `AnyType` provides explicit
downcasting for code that needs it. Anything that implements the `Type`
trait can be converted to `AnyType` via `as_any()`. The result may then
be used in a match expression to switch over all possible types. This is
useful for example in code generation or type checking, where different
code needs to run based on the type, but that code should not be part of
the `Type` trait itself.