vhdl: add floating point type
Add the `FloatingType` to the new `ty2` module, which represents ranges
of floating-point values. It's quite an awkward choice of VHDL to
allow floating-point values to be subtyped to a specific range. But
we'll support it anyway. We currently adhere to the standard in that we
represent floating-point values as Rust's `f64` type.