vhdl: add type marks
Add the `TypeMark` struct to the new `ty2` module. It is used to track
how the user referred to a type in the source code. It's more or less a
duplicate of `TypeDecl`, so there might be a possibility to unify the
two as one struct, with slightly different implementations of `Display`.