vhdl: add preliminary enum subtype
Add an initial `EnumSubtype` to the new `ty2`. It is based on a more
general `ScalarSubtype` which imposes a range constraint on any of the
scalar types. Type aliases exist for enum, integer, floating-point, and
physical types.

The subtype currently implements the `Subtype` trait, which does not do
much. I should probably find a way to unify types and subtypes under the
`Type` trait to facilitate ease of use and provide a common interface.