make moore fail properly on error diags
Currently the compiler's exit code merely depends on whatever result the
last operation invoked produced. This might hide errors that happened
while arriving at that result. For example, the compiler currently type
checks packages, which yields no result. Testing whether anything failed
requires observing the diagnostics as they fly by.

Extend the `Session` struct in `common` to keep track of whether any
diagnostic had error severity or worse, and fail with a non-zero exit
code if that was the case.

Disable some of the VHDL elaboration tests which keep on failing at the
moment because we're overhauling expressions and constant evaluation.