vhdl: impl first HIR slot loading
Add the `vhdl::debug` module to do some testing. Pass all VHDL design
units to this module when the `--emit-pkgs` option is passed to the
compiler. In the module, map all packages to slots, then ask each slot
for its span. This will lazily flatten the slot. Exactly what we need:
on-demand calculations.