vhdl: prepare for new scoping mechanism
Prepare the new HIR experiment for a scoping mechanism. Also add a
convenience `Result<T>` type with `()` as its error. Remove the
implementation of `Node` for `Slot` and add a more appropriate
`LatentNode` trait which represents the lazy loading of nodes.

Currently there is no way to access the details of a package's
declarations, since the type is erased to a `Node`. We need a mechanism
such that whatever slots lazily load can implement something more
specific than `Node`.