vhdl: factor out scoping into new module
Create a new `scope2` module which will someday replace `scope`. Add a
`ScopeData` struct which holds the definitions and imports made in a
scope, and a `ScopeContext` trait which abstracts interaction with a
scope. Remove the previous scoping business from `pkg.rs` and use the
new facilities instead. Implement `ScopeContext` for the `hir::Context`
struct, which makes declaring things a breeze. Also add a `subscope`
function which creates a new scope and returns a new context for that.