vhdl: start impl of type and const val calc
Start implementing the calculation of types and constant values for
ranges and expressions. This will pave the way for determining the
actual type of type declarations. This also requires figuring out how we
will do typeval/typeck with the new trait-based HIR nodes and contexts.
Pretty tricky; there probably needs to be a context trait that
encapsulates everything needed to act as a context fo typeval/typeck.