vhdl: impl int lit const values, integer types
Add the `konst2` module which shall contain an overhauled version of the
constant values facilities. Introduce a neat new `Cow`-based way of
conditionally allocating constants into an arena or reusing existing
references. Implement the calculation of constant values for integer
literals. Add the mapping of integer type declarations to instances of
`ty2::IntegerBasetype`. This now allows the `pkg_0.vhd` example to have
its types mapped properly.

There's lots to do, but the direction of `konst2` should be a good guide
for future adjustments to the `ty2` module. Also, the `ExprContext`
marker trait is proving to be very handy for expression manipulation.
Similat approaches can be taken for other things, e.g. constant or type
manipulation.