vhdl: impl access type declarations
Implement declaration of access types. When encountering an access type,
the subtype indication is mapped to a new `SubtypeInd2` which follows
the slot-based architecture. The declared type is determined by querying
the declared type of the subtype indication, and wrapping that in the
new `AccessType`. Note that `declared_type` for subtype indications is
not yet implemented and panics.