svlog: add param env
Add parameter environments. A `ParamEnv` encapsulates the values and
types assigned to parameters and interfaces during the instantiation of
a module. The environments are internalized into the global context and
used as second parameter to code generation. In the future also type
checking will have to carry this environment around.