svlog: add name resolution
Implement name resolution. Add the `Rib` struct and `RibKind` enum to
represent metadata about the scope associated with a node. Nodes may
then choose to generate their own ribs, e.g. port declarations, or not
do so. In the latter case, the node inherits its parent's rib. The ribs
thus form a tree that may be traversed in order to resolve a name.