svlog: allow const value/type in param binding
Extend parameter environments to allow for direct binding of value and
type parameters. That is, rather than referencing another node to query
for binding information, the environment can now carry a value or type
directly. This will be useful for the evaluation of for-generate loops
where we want to directly bind a value to a genvar.