codegen: add expr generator trait and structs
Add the `ExprGenerator` trait to the codegen module which implements the
code generation for expressions. Also add the `ProcessGenerator` struct
that wraps around an LLHD process being emitted. The latter implements
the former trait, providing an implementation of `emit_inst()`. This is
the first step towards allowing expressions to be emitted into entities.