codegen: differentiate rvalue emission modes
Add the `Mode` enum and `mode` parameter to `emit_rvalue` to control how
rvalues are emitted. This allows for calling code to specify if the
emitted value is expected to be a signal, pointer, or regular value.
This now allows for proper support of expressions in port assignments.
And it paves the way for prb/drv-less expressions in entities.