mir: add array pattern rvalue nodes
Add some very basic MIR rvalue nodes to represent `'{...}` patterns
being used as rvalue expressions. This is a first step towards #104.
The MIR allows for the types of expressions to be very precisely
specified and makes all cast operations explicit.

The nodes are currently all dummy nodes, but the code is in place which
would eventually assemble the MIR tree. No pretty printing yet.

Try:

    cargo run -- tests/svlog/mir/pattern.sv -e A -vvvv