mir: add int binary op, test for result size
Add a basic lowering implementation for integer binary arithmetic
operators. Also add a test that will allow for the result sizes to be
checked. The result is pretty incomplete, but is a good stepping stone
to weed out some type system inaccuracies and to start and get the
implicit casts to integer and related semantics right.