svlog: fix size cast parsing
Fix an issue in the lexer which would attempt to lex `32'(x)` as a sized
integer literal, but then realize that nothing like a literal follows
the `'`. In that case the lexer emits an apostrophe token, dropping the
`32`. Fix this by looking ahead and deciding whether what follows looks
like a literal.

Fixes #113.