svlog: ast: Add as_ptr() to AnyNode for id comp
This fixes a nasty bug where the query database would treat two calls to
a query with the same arguments as separate, if the argument was a dyn
reference, and the two had a different vtable pointer. The new function
casts to `*const u8`, which then allows for nodes to be compared purely
based on memory location.