svlog: Impl string literals
According to the SV manual, string literals are not of `string` type,
but rather behave like a packed array of bits that is `8*length` wide.
The array holds the ASCII characters of the string, with the first
character in the string appearing at the MSB end (such that assigning to
a variable that is too narrow would truncate the beginning of the
string).

Contributes towards #161.