mlir: Insert new blocks in-order after insertion point
Instead of just appending new blocks to the end of the region, actually
be clever about where to put those new blocks and insert them after the
current insertion point, or any block inserted after the last change to
the insertion point. This has the effect of adding blocks after any
explicitly set insertion point in the order they were created, which
makes for easy-to-read output.