hir: Add unmodified AST variant to StmtKind
Add a `StmtKind::Ast(&ast::Stmt)` variant which can be used to represent
unmodified AST nodes in the HIR. It is very likely that pretty much all
HIR nodes can be replaced with the corresponding AST node, and all the
relevant work can be done there. This is a solution to gradually perform
this transition.