codegen: Track whether block was terminated
Add a `terminated` flag to the `UnitGenerator` which gets set to true
when a branch or return is generated, and to false when the insertion
point is changed to a different block. This will help deal with
situations where the input SystemVerilog contains a terminator statement
like `return`, followed by additional statements that are unreachable.